+++ tests/testthat/test_balance.R
@@ -3,7 +3,7 @@ context("balance()")
 
 test_that("all size settings work in balance()",{
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   # Create data frame
   df <- data.frame(
@@ -12,17 +12,17 @@ test_that("all size settings work in balance()",{
     "score" = sample(c(1:100), 7))
 
   # Using balance() with number
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df_3 <- balance(df, 3, "participant")
 
   expect_equal(nrow(df_3), 3*3)
   expect_equal(df_3$participant, factor(c(1,1,1,2,2,2,3,3,3)))
-  expect_equal(df_3$trial, c(1,2,1,1,1,1,1,2,4))
+  expect_equal(df_3$trial, c(1,2,1,1,1,1,4,1,2))
   expect_equal(ncol(df_3), 3)
 
 
   # Using balance() with min
-  set.seed(2)
+  set_seed_for_R_compatibility(2)
   df_min <- balance(df, "min", "participant")
 
   expect_equal(nrow(df_min), 3)
@@ -31,16 +31,16 @@ test_that("all size settings work in balance()",{
   expect_equal(ncol(df_min), 3)
 
   # Using balance() with max
-  set.seed(2)
+  set_seed_for_R_compatibility(2)
   df_max <- balance(df, "max", "participant")
 
   expect_equal(nrow(df_max), 4*3)
   expect_equal(df_max$participant, factor(c(1,1,1,1,2,2,2,2,3,3,3,3)))
-  expect_equal(df_max$trial, c(1,2,1,1,1,1,1,1,1,2,3,4))
+  expect_equal(df_max$trial, c(1,2,1,2,1,1,1,1,1,2,3,4))
   expect_equal(ncol(df_max), 3)
 
   # Using balance() with mean
-  set.seed(19)
+  set_seed_for_R_compatibility(19)
   df_mean <- balance(df, "mean", "participant")
 
   expect_equal(nrow(df_mean), 2*3)
@@ -49,7 +49,7 @@ test_that("all size settings work in balance()",{
   expect_equal(ncol(df_mean), 3)
 
   # Using balance() with median
-  set.seed(19)
+  set_seed_for_R_compatibility(19)
   df_median <-balance(df, "median", "participant")
 
   expect_equal(nrow(df_median), 2*3)
@@ -62,7 +62,7 @@ test_that("all size settings work in balance()",{
 
 test_that("mark_new_rows works in balance()",{
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   # Create data frame
   df <- data.frame(
@@ -70,11 +70,11 @@ test_that("mark_new_rows works in balance()",{
     "trial" = c(1,2,1,1,2,3,4),
     "score" = sample(c(1:100), 7))
 
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df_3 <- balance(df, 3, "participant", mark_new_rows = TRUE)
   expect_equal(df_3$.new_row, c(0,0,1,0,1,1,0,0,0))
 
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df_3 <- balance(df, 3, "participant", mark_new_rows = TRUE, new_rows_col_name = "someName")
   expect_equal(df_3$someName, c(0,0,1,0,1,1,0,0,0))
 
@@ -82,7 +82,7 @@ test_that("mark_new_rows works in balance()",{
 
 test_that("both wrapper functions, upsample() and downsample() work",{
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   # Create data frame
   df <- data.frame(
@@ -90,16 +90,16 @@ test_that("both wrapper functions, upsample() and downsample() work",{
     "trial" = c(1,2,1,1,2,3,4),
     "score" = sample(c(1:100), 7))
 
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df_min_balance <- balance(df, "min", "participant")
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df_min_downsample <- downsample(df, "participant")
 
   expect_equal(df_min_balance, df_min_downsample)
 
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df_max_balance <- balance(df, "max", "participant")
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df_max_upsample <- upsample(df, "participant")
 
   expect_equal(df_max_balance, df_max_upsample)
@@ -108,11 +108,11 @@ test_that("both wrapper functions, upsample() and downsample() work",{
 
 test_that("balance() works in dplyr pipeline",{
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   library(dplyr)
   # Create data frame
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df <- data.frame(
     "participant" = factor(c(1, 1, 2, 3, 3, 3, 3)),
     "trial" = c(1,2,1,1,2,3,4),