+++ tests/testthat/test_group.R
@@ -5,7 +5,7 @@ context("group()")
 
 test_that("dimensions of data frame with group()", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   df <- data.frame("x"=c(1:12),
                    "species" = rep(c('cat','pig', 'human'), 4),
@@ -26,7 +26,7 @@ test_that("dimensions of data frame with group()", {
 
 test_that("mean age of groups made with group()", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   # Create df 3x12
   df <- data.frame("x"=c(1:12),
@@ -60,7 +60,7 @@ test_that("mean age of groups made with group()", {
 
 test_that("error messages work in group()", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   # Create df 3x12
   df <- data.frame("x"=c(1:12),
@@ -74,7 +74,7 @@ test_that("error messages work in group()", {
 
 test_that("allow_zero works in group()", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   # Create df 3x12
   df <- data.frame("x"=c(1:12),
@@ -114,7 +114,7 @@ test_that("allow_zero works in group()", {
 
 test_that("col_name can be set correctly in group()",{
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   # Create df 3x12
   df <- data.frame("x"=c(1:12),
@@ -135,7 +135,7 @@ test_that("col_name can be set correctly in group()",{
 
 test_that("l_starts can take n = auto", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   df <- data.frame("x"=c(1:12),
                    "species" = rep(c('cat','cat', 'human', 'human'), 3),
@@ -156,7 +156,7 @@ test_that("l_starts can take n = auto", {
 
 test_that("l_starts can take starts_col = index / .index", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   df <- data.frame("x"=c(1:12))