+++ tests/testthat/test_grouping_factor.R
@@ -4,7 +4,7 @@ context("group_factor()")
 
 test_that("group sizes works with group_factor with method greedy", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   group_counts <- function(v, n){
 
@@ -41,7 +41,7 @@ test_that("group sizes works with group_factor with method greedy", {
 
 test_that("group sizes works with group_factor with method n_dist", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   group_counts <- function(v, n){
 
@@ -78,7 +78,7 @@ test_that("group sizes works with group_factor with method n_dist", {
 
 test_that("group sizes works with group_factor with method n_fill", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   group_counts <- function(v, n){
 
@@ -117,7 +117,7 @@ test_that("group sizes works with group_factor with method n_fill", {
 
 test_that("group sizes works with group_factor with method n_rand", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   count_groups <- function(v, n){
 
@@ -167,7 +167,7 @@ test_that("group sizes works with group_factor with method n_rand", {
 
 test_that("group sizes works with group_factor with method staircase", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   group_counts <- function(v, n){
 
@@ -235,7 +235,7 @@ test_that("group sizes works with group_factor with method staircase", {
 
 test_that("group sizes works with group_factor with method primes", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   group_counts <- function(v, n){
 
@@ -293,7 +293,7 @@ test_that("group sizes works with group_factor with method primes", {
 
 test_that("group sizes works with group_factor with method l_sizes", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   group_f <- function(v, n){
 
@@ -325,7 +325,7 @@ test_that("group sizes works with group_factor with method l_sizes", {
 
 test_that("group sizes works with group_factor with method l_starts", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   group_f <- function(v, n){
 
@@ -400,7 +400,7 @@ test_that("group sizes works with group_factor with method l_starts", {
 
 test_that("l_starts raises error correctly when value is not found", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   expect_error(group(c(1:5), c(1,3,6), method = 'l_starts'),
                "Start value \"6\" not found in vector", fixed = TRUE)
@@ -425,7 +425,7 @@ test_that("l_starts raises error correctly when value is not found", {
 
 test_that("force_equal works with group_factor with all methods",{
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   group_counts <- function(v, n, method){
 
@@ -508,7 +508,7 @@ test_that("force_equal works with group_factor with all methods",{
 
 test_that("allow_zero works with group_factor", {
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   v = c(1:3)
 
@@ -520,7 +520,7 @@ test_that("allow_zero works with group_factor", {
 
 test_that("l_sizes works with no groups",{
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   expect_equal(group_factor(c(1:3), 0.2, method = 'l_sizes'), factor(c(2,2,2)))
   expect_warning(group_factor(c(1:3), 0.2, method = 'l_sizes', force_equal = TRUE), "No groups. Returned NA.", fixed = TRUE)