+++ tests/testthat/test_numerically_balanced_grouping.R
@@ -3,10 +3,8 @@ context("numerically_balanced_group_factor_()")
 
 test_that("numerically_balanced_group_factor_() work with n=2", {
 
-  skip_test_if_old_R_version()
-
   # Create data frame
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df <- data.frame(
     "participant" = factor(c(1, 3, 5, 6, 7, 8)),
     "score" = c(79,85,140,69,87,92))
@@ -18,17 +16,15 @@ test_that("numerically_balanced_group_factor_() work with n=2", {
     dplyr::group_by(.groups) %>%
     dplyr::summarize(group_sum = sum(score))
 
-  expect_equal(vbf, factor(c(2,1,1,2,1,2)))
-  expect_equal(group_sums$group_sum, c(312,240))
+  expect_equal(vbf, factor(c(1,2,2,1,2,1)))
+  expect_equal(group_sums$group_sum, c(240,312))
 
 })
 
 test_that("numerically_balanced_group_factor_() works with n=3", {
 
-  skip_test_if_old_R_version()
-
   # Create data frame
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df <- data.frame(
     "participant" = factor(c(1, 1, 2, 3, 3, 3, 3)),
     "trial" = c(1,2,1,1,2,3,4),
@@ -37,19 +33,19 @@ test_that("numerically_balanced_group_factor_() works with n=3", {
                   neg_pos_score = score-50)
 
   # numerically_balanced_group_factor_ on unequal number of data frame rows
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   expect_equal(numerically_balanced_group_factor_(df, 3, num_col="score"),
-               factor(c(2,3,3,2,1,3,1)))
+               factor(c(1,3,2,3,1,2,1)))
 
 
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   nbf1 <- numerically_balanced_group_factor_(df, 3, num_col="neg_score")
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   nbf2 <- numerically_balanced_group_factor_(df, 3, num_col="score")
   expect_equal(nbf1,nbf2)
 
   # add grouping factor to df and get sums of value col
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df_grouped <- df %>%
     dplyr::mutate(.groups = numerically_balanced_group_factor_(df, 3, num_col="score"))
 
@@ -58,26 +54,26 @@ test_that("numerically_balanced_group_factor_() works with n=3", {
     dplyr::summarize(group_sum = sum(score))
   # group_sums
 
-  expect_equal(group_sums$group_sum, c(101,102,83))
+  expect_equal(group_sums$group_sum, c(144,143,126))
 
   # numerically_balanced_group_factor_ on equal number of data frame rows
 
   df <- df %>% dplyr::filter(dplyr::row_number() != 7)
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   expect_equal(numerically_balanced_group_factor_(df, 3, num_col="score"),
-               factor(c(2,3,1,2,1,3)))
+               factor(c(3,2,2,1,1,3)))
 
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   nbf1 <- numerically_balanced_group_factor_(df, 3, num_col="neg_score")
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   nbf2 <- numerically_balanced_group_factor_(df, 3, num_col="score")
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   nbf3 <- numerically_balanced_group_factor_(df, 3, num_col="neg_pos_score")
   expect_equal(nbf1,nbf2)
   expect_equal(nbf1,nbf3)
 
   # add grouping factor to df and get sums of value col
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df_grouped <- df %>%
     dplyr::mutate(.groups = numerically_balanced_group_factor_(
       df, 3, num_col="score"))
@@ -86,7 +82,7 @@ test_that("numerically_balanced_group_factor_() works with n=3", {
     dplyr::group_by(.groups) %>%
     dplyr::summarize(group_sum = sum(score))
 
-  expect_equal(group_sums$group_sum, c(88,102,82))
+  expect_equal(group_sums$group_sum, c(109,94,113))
 
 })
 
@@ -108,7 +104,7 @@ test_that("numerically_balanced_group_factor_() unequal method on small datasets
 
     group_summaries <- plyr::ldply(1:iter, function(i){
 
-      set.seed(i)
+      set_seed_for_R_compatibility(i)
       if (i %% 2 == 0){
         cond <- "uniform"
         df$score <- sample(1:100, n_rows)
@@ -122,7 +118,6 @@ test_that("numerically_balanced_group_factor_() unequal method on small datasets
       df_grouped <- df %>%
         dplyr::mutate(.groups = numerically_balanced_group_factor_(., n_folds,
                                                                    num_col="score",
-                                                                   randomize_pairs = T,
                                                                    unequal_method=unequal_method),
                       condition = cond)
 
@@ -231,10 +226,8 @@ test_that("numerically_balanced_group_factor_() unequal method on small datasets
 
 test_that("numerically_balanced_group_factor_() work method='l_sizes'", {
 
-  skip_test_if_old_R_version()
-
   # Create data frame
-  set.seed(1)
+  set_seed_for_R_compatibility(2)
   df <- data.frame(
     "participant" = factor(c(1, 3, 5, 6, 7, 8)),
     "score" = c(79,85,140,69,87,92))
@@ -250,6 +243,7 @@ test_that("numerically_balanced_group_factor_() work method='l_sizes'", {
   expect_equal(vbf, factor(c(1,2,1,1,2,2)))
   expect_equal(group_sums$group_sum, c(288,264))
 
+  set_seed_for_R_compatibility(4)
   vbf <- numerically_balanced_group_factor_(df, 0.2, num_col="score", method = 'l_sizes')
   df_vbf <- df %>%
     dplyr::mutate(.groups = vbf)
@@ -258,9 +252,10 @@ test_that("numerically_balanced_group_factor_() work method='l_sizes'", {
     dplyr::group_by(.groups) %>%
     dplyr::summarize(group_sum = sum(score))
 
-  expect_equal(vbf, factor(c(2,2,2,1,2,2)))
-  expect_equal(group_sums$group_sum, c(69,483))
+  expect_equal(vbf, factor(c(1,2,2,2,2,2)))
+  expect_equal(group_sums$group_sum, c(79,473))
 
+  set_seed_for_R_compatibility(19)
   # With 3 partitions
   vbf <- numerically_balanced_group_factor_(df, c(0.2,0.2), num_col="score",
                                             method = 'l_sizes')
@@ -271,20 +266,18 @@ test_that("numerically_balanced_group_factor_() work method='l_sizes'", {
     dplyr::group_by(.groups) %>%
     dplyr::summarize(group_sum = sum(score))
 
-  expect_equal(vbf, factor(c(3,1,3,3,2,3)))
-  expect_equal(group_sums$group_sum, c(85,87,380))
+  expect_equal(vbf, factor(c(3,3,2,1,3,3)))
+  expect_equal(group_sums$group_sum, c(69,140,343))
 
 })
 
 
 test_that("numerically_balanced_group_factor_() on large datasets", {
 
-  skip_test_if_old_R_version()
-
   # testthat::skip(message = "Skipping numerical balancing of a large dataset")
 
   # Create data frame
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df <- data.frame(
     "participant" = factor(rep(1:10, 100)),
     "diagnosis" = factor(rep(rep(c("a","b"), each=5), 100)),
@@ -297,11 +290,11 @@ test_that("numerically_balanced_group_factor_() on large datasets", {
     dplyr::group_by(num_balanced_factor_1) %>%
     dplyr::summarize(group_sum = sum(score))
 
-  expect_equal(group_sums$group_sum, c(100,100,99.8,99.9,99.9), tolerance=1e-3)
+  expect_equal(group_sums$group_sum, c(100,100,99.9,99.7,100), tolerance=1e-3)
 
   ####
 
-  set.seed(1)
+  set_seed_for_R_compatibility(6)
   df_999 <- head(df, 999)
 
   df_999$num_balanced_factor_2 <- numerically_balanced_group_factor_(df_999,
@@ -314,10 +307,8 @@ test_that("numerically_balanced_group_factor_() on large datasets", {
     dplyr::summarize(group_sum = sum(score),
                      group_count = dplyr::n())
 
-  expect_equal(group_summaries$group_sum, c(99.9,100,100,98.9,100), tolerance=1e-2)
-  expect_equal(group_summaries$group_count, c(200,200,200,199,200), tolerance=1e-2)
+  expect_equal(group_summaries$group_sum, c(99.95,100.15,100.05,99.01,100.24), tolerance=1e-4)
+  expect_equal(group_summaries$group_count, c(200,200,200,199,200), tolerance=1e-3)
 
 
 })
-
-