+++ tests/testthat/test_partition.R
@@ -3,7 +3,7 @@ context("partition()")
 
 test_that("dimensions of data frame with partition()",{
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   df <- data.frame("participant" = factor(rep(c('1','2', '3', '4', '5', '6'), 3)),
                    "age" = rep(c(25,65,34), 3),
@@ -28,7 +28,7 @@ test_that("dimensions of data frame with partition()",{
 
 test_that(".partitions is correct in partition() with list_out == FALSE",{
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(6)
 
   df <- data.frame("participant" = factor(rep(c('1','2', '3', '4', '5', '6'), 3)),
                    "age" = rep(c(25,65,34), 3),
@@ -45,7 +45,7 @@ test_that(".partitions is correct in partition() with list_out == FALSE",{
 
   col_is_factor <- function(df, n, cat_col = NULL, num_col = NULL, id_col = NULL, col){
 
-    set.seed(1)
+    set_seed_for_R_compatibility(1)
     partitioned_df <- partition(df, n, cat_col=cat_col, num_col=num_col,
                                 id_col=id_col, list_out = FALSE)
 
@@ -57,7 +57,7 @@ test_that(".partitions is correct in partition() with list_out == FALSE",{
 
   group_counts <- function(df, n, cat_col = NULL, num_col = NULL, id_col = NULL, force_equal = FALSE){
 
-    set.seed(1)
+    set_seed_for_R_compatibility(1)
     partitioned_df <- partition(df, n, cat_col=cat_col, num_col=num_col,
                                 id_col=id_col, force_equal = force_equal,
                                 list_out = FALSE)
@@ -106,7 +106,7 @@ test_that(".partitions is correct in partition() with list_out == FALSE",{
   expect_equal(group_counts(df, 1, cat_col = 'diagnosis', id_col = 'participant', num_col = 'score'), c(6,12))
 
   # unequal number of rows in df
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   expect_equal(group_counts(df_unequal, 0.2), c(3,14))
   expect_equal(group_counts(df_unequal, 0.2, cat_col = 'diagnosis'), c(2,15))
   expect_equal(group_counts(df_unequal, 0.2, id_col = 'participant'), c(3,14))
@@ -115,11 +115,11 @@ test_that(".partitions is correct in partition() with list_out == FALSE",{
                             id_col = 'participant'), c(6,11))
   expect_equal(group_counts(df_unequal, 0.4, cat_col = 'diagnosis',
                             num_col = 'score'), c(6,11))
-  expect_equal(group_counts(df_unequal, 0.4, num_col = 'score', id_col = 'participant'), c(5,12))
+  expect_equal(group_counts(df_unequal, 0.4, num_col = 'score', id_col = 'participant'), c(6,11))
   expect_equal(group_counts(df_unequal, 0.4, cat_col = 'diagnosis', num_col = 'score',
-                            id_col = 'participant'), c(5,12))
+                            id_col = 'participant'), c(6,11))
   expect_equal(group_counts(df_unequal, 2, cat_col = 'diagnosis',
-                            id_col = 'participant'), c(12,5))
+                            id_col = 'participant'), c(11,6))
   expect_equal(group_counts(df_unequal, 3, cat_col = 'diagnosis', id_col = 'participant'), c(17))
   expect_equal(group_counts(df_unequal, 3, id_col = 'participant'), c(9,8))
   expect_equal(group_counts(df_unequal, 2, id_col = 'participant'), c(6,11))
@@ -127,15 +127,15 @@ test_that(".partitions is correct in partition() with list_out == FALSE",{
   expect_equal(group_counts(df_unequal, c(2,4,6), num_col = 'score'), c(2,4,6,5))
   expect_equal(group_counts(df_unequal, 2, cat_col = 'diagnosis', num_col = 'score'), c(4,13))
   expect_equal(group_counts(df_unequal, 2, cat_col = 'diagnosis', num_col = 'score'), c(4,13))
-  expect_equal(group_counts(df_unequal, 2, id_col = 'participant', num_col = 'score'), c(5,12))
+  expect_equal(group_counts(df_unequal, 2, id_col = 'participant', num_col = 'score'), c(6,11))
   expect_equal(group_counts(df_unequal, 2, cat_col = 'diagnosis', id_col = 'participant', num_col = 'score'), c(11,6))
-  expect_equal(group_counts(df_unequal, 1, cat_col = 'diagnosis', id_col = 'participant', num_col = 'score'), c(5,12))
+  expect_equal(group_counts(df_unequal, 1, cat_col = 'diagnosis', id_col = 'participant', num_col = 'score'), c(6,11))
 
 
   # Test force_equal
 
   # equal number of rows in df
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   expect_equal(group_counts(df, 2, id_col = 'participant', force_equal = TRUE), c(6))
   expect_equal(group_counts(df, 3, id_col = 'participant', force_equal = TRUE), c(9))
   expect_equal(group_counts(df, 2, cat_col = 'diagnosis', force_equal = TRUE), c(4))
@@ -148,7 +148,7 @@ test_that(".partitions is correct in partition() with list_out == FALSE",{
                             force_equal = TRUE), c(12))
 
   # unequal number of rows in df
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   expect_equal(group_counts(df_unequal, 2, id_col = 'participant', force_equal = TRUE), c(6))
   expect_equal(group_counts(df_unequal, 3, id_col = 'participant', force_equal = TRUE), c(9))
   expect_equal(group_counts(df_unequal, 2, cat_col = 'diagnosis', force_equal = TRUE), c(4))
@@ -164,7 +164,7 @@ test_that(".partitions is correct in partition() with list_out == FALSE",{
 
 test_that(".partitions is correct in partition() with list_out == TRUE",{
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   df <- data.frame("participant" = factor(rep(c('1','2', '3', '4', '5', '6'), 3)),
                    "age" = rep(c(25,65,34), 3),
@@ -216,7 +216,7 @@ test_that(".partitions is correct in partition() with list_out == TRUE",{
 
 test_that("partition() outputs correct error messages",{
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   df <- data.frame("participant" = factor(rep(c('1','2', '3', '4', '5', '6'), 3)),
                    "age" = rep(c(25,65,34), 3),
@@ -242,7 +242,7 @@ test_that("partition() outputs correct error messages",{
 
 test_that("bootstrap test of num_col works",{
 
-  skip_test_if_old_R_version()
+  set_seed_for_R_compatibility(1)
 
   # Takes 4 seconds, so we disable it for now.
   testthat::skip(message = "Skipping bootstrapped numerical balancing test in partition()")
@@ -252,7 +252,7 @@ test_that("bootstrap test of num_col works",{
                    "age"=rep(sample(100),100))
 
   # Single
-  set.seed(1)
+  set_seed_for_R_compatibility(1)
   df_partitioned <- partition(df, c(0.3,0.3),
                               #cat_col="diagnosis",
                               num_col="age",
@@ -261,7 +261,7 @@ test_that("bootstrap test of num_col works",{
 
   for (i in 1:10){
 
-    set.seed(i)
+    set_seed_for_R_compatibility(i)
     df_partitioned <- partition(df, 0.5, cat_col="diagnosis", num_col="age",
                                 id_col="participant", list_out = FALSE)
 
@@ -269,14 +269,14 @@ test_that("bootstrap test of num_col works",{
       dplyr::summarise(mean_age = mean(age),
                        sd_age = sd(age))
 
-    expect_true(is_between_(age_distribution$mean_age[1], 50, 52))
-    expect_true(is_between_(age_distribution$mean_age[2], 49, 51))
+    expect_true(is_between_(age_distribution$mean_age[1], 49, 53))
+    expect_true(is_between_(age_distribution$mean_age[2], 49, 53))
 
   }
 
   for (i in 1:10){
 
-    set.seed(i)
+    set_seed_for_R_compatibility(i)
     df_partitioned <- partition(df, c(0.2, 0.2, 0.2, 0.2, 0.2),
                                 cat_col="diagnosis", num_col="age",
                                 id_col="participant", list_out = FALSE)
@@ -297,7 +297,7 @@ test_that("bootstrap test of num_col works",{
 
   for (i in 1:10){
 
-    set.seed(i)
+    set_seed_for_R_compatibility(i)
     df_partitioned <- partition(df, c(0.2, 0.2, 0.2, 0.2, 0.2),
                                 cat_col="diagnosis", num_col="age",
                                 id_col="participant", extreme_pairing_levels = 2,
@@ -319,7 +319,7 @@ test_that("bootstrap test of num_col works",{
 
   for (i in 1:10){
 
-    set.seed(i)
+    set_seed_for_R_compatibility(i)
     df_partitioned <- partition(df, c(0.2, 0.2, 0.2, 0.2, 0.2),
                                 cat_col="diagnosis", num_col="age",
                                 id_col="participant", extreme_pairing_levels = 3,
@@ -329,11 +329,11 @@ test_that("bootstrap test of num_col works",{
       dplyr::summarise(mean_age = mean(age),
                        sd_age = sd(age))
 
-    expect_true(is_between_(age_distribution$mean_age[1], 49, 51.5))
-    expect_true(is_between_(age_distribution$mean_age[2], 49, 51.5))
-    expect_true(is_between_(age_distribution$mean_age[3], 49, 51.5))
-    expect_true(is_between_(age_distribution$mean_age[4], 49, 51.5))
-    expect_true(is_between_(age_distribution$mean_age[5], 49, 51.5))
+    expect_true(is_between_(age_distribution$mean_age[1], 49, 53))
+    expect_true(is_between_(age_distribution$mean_age[2], 49, 53))
+    expect_true(is_between_(age_distribution$mean_age[3], 49, 53))
+    expect_true(is_between_(age_distribution$mean_age[4], 49, 53))
+    expect_true(is_between_(age_distribution$mean_age[5], 49, 53))
 
   }